<feat> <combination> : <完成资产配置-目标+路径和资产配置-基金+组合>

+++ src/main/java/reaper/bean/AssetWeightBean.java
@@ -9,19 +9,19 @@ public class AssetWeightBean {
     /**
      * 股票型基金
      */
-    public Integer stock;
+    public Double stock;
 
     /**
      * 债券型基金
      */
-    public Integer bond;
+    public Double bond;
 
     /**
      * 混合型基金
      */
-    public Integer hybrid;
+    public Double hybrid;
 
-    public AssetWeightBean(Integer stock, Integer bond, Integer hybrid) {
+    public AssetWeightBean(Double stock, Double bond, Double hybrid) {
         this.stock = stock;
         this.bond = bond;
         this.hybrid = hybrid;
